Adjusting table creation fields

+++ cg-rails/db/migrate/20230501203125_create_prompts.rb
@@ -1,7 +1,7 @@
 class CreatePrompts < ActiveRecord::Migration[7.0]
   def change
     create_table :prompts do |t|
-      t.integer :original_index, null: false, default: 0
+      t.bigint :original_index, null: false, default: 0
       t.text :content
       t.timestamps
     end
